refactor: avoid resolving logos on the fly
