+++ packages/launchpad/package.json
@@ -33,7 +33,7 @@
     "vite-plugin-components": "0.11.3",
     "vite-plugin-icons": "0.6.3",
     "vite-plugin-windicss": "1.1.1",
-    "vue-prism-component": "^2.0.0",
+    "vue-prism-component": "2.0.0",
     "vue-tsc": "0.0.24",
     "windicss": "3.1.3"
   },
